// frontend/src/types.ts
export type ManufacturingUnit = {
  id: number
  batch_code: string
  company_name: string
  address: string
  fssai_licence_number: string
  sort_order: number
}

export type UnitFormData = {
  batch_code: string
  company_name: string
  address: string
  fssai_licence_number: string
  sort_order: number
}